+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### Next
+
+* `MSeq`
+  * New functions: `sliceSummaryMay`, `sliceSummary`,
+    `foldlSliceSummaryComponents`.
+
 ### 0.1.0.0 -- 2024-06-16
 
 * First version.